+++ src/constant/GOOGLE_CONSTANT.ts
@@ -2,7 +2,7 @@ const GOOGLE_CONSTANT = {
   url: "https://www.google.com/search",
   inverse: {
     base: "https://lens.google.com/",
-    upload: "https://lens.google.com/upload",
+    upload: "https://lens.google.com/v3/upload",
     url: "https://lens.google.com/uploadbyurl?url=",
   },
   queryParam: "tbs",
@@ -11,7 +11,7 @@ const GOOGLE_CONSTANT = {
   },
   headers: {
     "User-Agent":
-      "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/103.0.0.0 Safari/537.36",
+      "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/121.0.0.0 Safari/537.36 Edg/121.0.0.0",
   },
 };