+++ src/core/parseGoogleImages.ts
@@ -1,28 +1,20 @@
-import { isImage, unicodeToChar } from "../utils/utils";
+import { unicodeToChar } from "../utils/utils";
 import GOOGLE_CONSTANT from "../constant/GOOGLE_CONSTANT";
 import axios, { AxiosProxyConfig } from "axios";
 import ImageResultItem from "../types/imageResultItem";
-
-const { FastHTMLParser } = require("fast-html-dom-parser");
 
 /**
  * Scrap google images scripts tag
  * @param url
  * @returns
  */
-async function scrapGoogleImagesScriptsTag(
-  url: string,
-  proxy?: AxiosProxyConfig
-) {
+async function scrapGoogleImages(url: string, proxy?: AxiosProxyConfig) {
   const { data } = await axios(url, {
     headers: GOOGLE_CONSTANT.headers,
     ...(proxy ?? {}),
   });
 
-  const parser = new FastHTMLParser(data);
-  const scripts = parser.getElementsByTagName("script");
-
-  return scripts;
+  return data;
 }
 
 /**
@@ -35,23 +27,11 @@ function getGoogleImageObject(
   informationsMatch: RegExpExecArray,
   otherInformationsMatch: RegExpExecArray
 ) {
-  const [r, g, b] = [
-    informationsMatch[4],
-    informationsMatch[5],
-    informationsMatch[6],
-  ].map((e) => parseInt(e, 10));
-
   return {
     id: otherInformationsMatch[1],
     title: otherInformationsMatch[3],
     url: unicodeToChar(informationsMatch[1]),
     originalUrl: otherInformationsMatch[2],
-    averageColor: `rgb(${r}, ${g}, ${b})`,
-    averageColorObject: {
-      r,
-      g,
-      b,
-    },
     height: parseInt(informationsMatch[2], 10),
     width: parseInt(informationsMatch[3], 10),
   };
@@ -68,45 +48,35 @@ async function parseGoogleImages(
 ): Promise<ImageResultItem[]> {
   const result: ImageResultItem[] = [];
 
-  const scripts = await scrapGoogleImagesScriptsTag(url, proxy);
+  const body: string = await scrapGoogleImages(url, proxy);
 
-  if (!scripts) return result;
+  //getting image url, height, width, color average
+  const informationsRegex = /\["(http[^"]+?)",(\d+),(\d+)\]/gi;
+  //getting originalUrl, title, id
+  const otherInformationsRegex =
+    /\[[\w\d]+?,"([^"]+?)","(http[^"]+?)","([^"]+?)"/gi;
 
-  for (const script of scripts) {
-    const body = script.innerHTML;
+  let informationsMatch: RegExpExecArray;
 
-    // if we dont find any image extension we can skip
-    if (!isImage(body)) continue;
+  while ((informationsMatch = informationsRegex.exec(body)) !== null) {
+    if (informationsMatch[1].startsWith("https://encrypted-tbn0.gstatic.com"))
+      continue;
 
-    //getting image url, height, width, color average
-    const informationsRegex =
-      /\["(http[^"]+?)",(\d+),(\d+)\],[\w\d]+?,[\w\d]+?,"rgb\((\d+),(\d+),(\d+)\)"/gi;
-    //getting originalUrl, title, id
-    const otherInformationsRegex =
-      /\[[\w\d]+?,"([^"]+?)","(http[^"]+?)","([^"]+?)"/gi;
+    const otherInformationsMatch = otherInformationsRegex.exec(body);
 
-    let informationsMatch: RegExpExecArray,
-      otherInformationsMatch: RegExpExecArray;
+    if (otherInformationsMatch === null) return result;
 
-    while (
-      (informationsMatch = informationsRegex.exec(body)) !== null &&
-      (otherInformationsMatch = otherInformationsRegex.exec(body)) !== null
-    ) {
-      if (informationsMatch.length < 4 || otherInformationsMatch.length < 4)
-        continue;
-      if (
-        informationsMatch[1].match(/http/gi).length > 2 ||
-        otherInformationsMatch[2].match(/http/gi).length > 2
-      )
-        continue;
+    if (informationsMatch.length < 4 || otherInformationsMatch.length < 4)
+      continue;
+    if (
+      informationsMatch[1].match(/http/gi).length > 2 ||
+      otherInformationsMatch[2].match(/http/gi).length > 2
+    )
+      continue;
 
-      result.push(
-        getGoogleImageObject(informationsMatch, otherInformationsMatch)
-      );
-    }
-
-    //if we get the correct scripts with all images we can exit
-    if (result.length > 0) return result;
+    result.push(
+      getGoogleImageObject(informationsMatch, otherInformationsMatch)
+    );
   }
 
   return result;