+++ src/types/imageResultItem.ts
@@ -3,12 +3,6 @@ type ImageResultItem = {
   title: string;
   originalUrl: string;
   url: string;
-  averageColor: string;
-  averageColorObject: {
-    r: number;
-    g: number;
-    b: number;
-  };
   height: number;
   width: number;
 };